# specs/001-ai-security-reviewer-spec/contracts/api.openapi.yaml
openapi: 3.0.3
info:
  title: AI Security Reviewer API
  version: 0.1.0
  description: API contract overview for review orchestration, progress streaming, findings, history, export, and ownership verification.
servers:
  - url: https://api.dev.ai-security-reviewer.example
security:
  - bearerAuth: []
paths:
...
  /findings/{findingId}/status:
    patch:
      summary: Update finding in-session status
      parameters:
        - name: findingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolutionState]
              properties:
                resolutionState:
                  type: string
                  enum: [open, resolved]
      responses:
        '200':
          description: Updated
...
  /exports:
    post:
      summary: Request report export artifact generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '202':
          description: Export job accepted
  /exports/{exportJobId}:
    get:
      summary: Get export job status/result
      parameters:
        - name: exportJobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Export job status
...
components:
...
  schemas:
...
    ExportRequest:
      type: object
      required: [reviewSessionId, format, sections]
      properties:
        reviewSessionId:
          type: string
        format:
          type: string
          enum: [markdown, pdf, json]
        sections:
          type: array
          items:
            type: string
            enum: [summary, perspectiveRatings, findings, remediationCode, promptHistory]
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT